Moved actions into static file
